Add support for ShiftJIS art and /qst/ colored text (#1)
* Add ShiftJIS styles and BBcode support

For the FoolFuuka theme, the styles were moved out of the `.theme_default` block
so that they work with the Midnight style too. Also, the fallback font was
changed to monospace from sans serif.

* Support /qst/ colored text

* Bump version numbers

* Add white-space: pre to .shift-jis

* Add rest of ShiftJIS CSS, remove !important

Without the overflow CSS, the art wraps on mobile and looks bad. line-height
makes it more compact. Also, !important's weren't needed.

* Don't overflow heuristic SJIS art